refactor: view 메서드 이름 변경, 댓글 개수 countBy 사용

+++ src/main/java/kr/ac/project/boardproject/dto/response/BoardResponseDto.java
@@ -15,6 +15,6 @@ public class BoardResponseDto {
     private String title;
     private String content;
     private Long hit;
-    private int replyCount;
+    private Long replyCount;
     private LocalDateTime registerDate;
 }
